refactor: Improve readability and naming of actions
