+++ src/utils/parseUtils.js
@@ -1,5 +1,6 @@
 
 export const getPosts = (response) => response.data.children.reduce(postReducer, [])
+
 const postReducer = (posts, child) => {
 
     const post = child.data
@@ -7,6 +8,7 @@ const postReducer = (posts, child) => {
     return posts
 
 }
+
 export const isImage = (post) => post.preview ? true : false
 export const isStickied = (post) => post.stickied ? true : false
 export const getId = (post) => post.id
@@ -14,5 +16,27 @@ export const getTitle = (post) => post.title
 export const isGif = (post) => post.preview && post.preview.variants && post.preview.variants.gif ? true : false
 export const getUrl = (post) => isGif(post) ? post.preview.variants.gif.source.url : post.preview.images[0].source.url
 
-export const generateImageData = (post) => ({id: getId(post), url: getUrl(post), title: getTitle(post), gif: isGif(post)})
+export const getPostImageInfo = (post) => ({id: getId(post), url: getUrl(post), title: getTitle(post), gif: isGif(post)})
 
+export const extractForStore = (post) => {
+
+    const images = []
+    const likes = {}
+    const nsfw = {}
+    const gifDuration = {}
+
+    posts.map((post) => {
+
+        const imageData = getPostImageInfo(post)
+        images.push(imageData)
+
+        const {id} = imageData
+        likes[id] = 0
+        nsfw[id] = false
+        gifDuration[id] = 0
+
+    })
+
+    return {images, likes, nsfw, gifDuration}
+
+}
